requirements.txt
absl-py==2.2.1
asgiref==3.8.1
astunparse==1.6.3
build==1.2.2.post1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
Django==5.1.7
djangorestframework==3.15.2
flatbuffers==25.2.10
gast==0.6.0
google-pasta==0.2.0
grpcio==1.71.0
h5py==3.13.0
idna==3.10
imageio==2.37.0
joblib==1.4.2
keras==3.9.1
lazy_loader==0.4
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml_dtypes==0.5.1
namex==0.0.8
networkx==3.4.2
numpy==2.1.3
nvidia-nccl-cu12==2.26.2
opencv-python==4.11.0.86
opt_einsum==3.4.0
optree==0.14.1
packaging==24.2
pillow==11.1.0
pip-tools==7.4.1
protobuf==5.29.4
Pygments==2.19.1
pyproject_hooks==1.2.0
requests==2.32.3
rich==13.9.4
scikit-image==0.25.2
scipy==1.15.2
six==1.17.0
sqlparse==0.5.3
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
tifffile==2025.3.13
typing_extensions==4.13.0
urllib3==2.3.0
Werkzeug==3.1.3
wrapt==1.17.2
xgboost==3.0.0
